initial commit of async branch.
This branch will contain the work to optionally use the library
asynchronously.

All api communication and deserialization is now in a class
Endpoints in the api module. All api.get calls have been replaced
with calls to the new endpoints. Also all the endpoints for the
convore api are already available.